+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/properties/events/DamageType.java
@@ -1,0 +1,11 @@
+package insane96mcp.mobspropertiesrandomness.data.json.properties.events;
+
+
+import com.google.gson.annotations.SerializedName;
+
+public enum DamageType {
+    @SerializedName("direct")
+    DIRECT,
+    @SerializedName("indirect")
+    INDIRECT
+}