+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/properties/events/Target.java
@@ -1,0 +1,11 @@
+package insane96mcp.mobspropertiesrandomness.data.json.properties.events;
+
+
+import com.google.gson.annotations.SerializedName;
+
+public enum Target {
+    @SerializedName("entity")
+    ENTITY,
+    @SerializedName("other")
+    OTHER
+}