+++ src/main/java/insane96mcp/mobspropertiesrandomness/setup/Strings.java
@@ -5,6 +5,8 @@ import insane96mcp.mobspropertiesrandomness.MobsPropertiesRandomness;
 public class Strings {
 	public static class Tags {
 		public static final String PROCESSED = MobsPropertiesRandomness.RESOURCE_PREFIX + "processed";
-		public static final String ON_HIT_EFFECTS = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_hit_effects";
+		public static final String ON_ATTACK = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_attack";
+		public static final String ON_ATTACKED = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_attacked";
+		public static final String ON_DEATH = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_death";
 	}
 }